Added basic socket frame.. still learn fastapi
