+++ templates/siteinfo.html
@@ -4,7 +4,7 @@
 <head>
     <meta charset="utf-8" />
     <title>Dashboard | Main Page</title>
-    <link rel="stylesheet" type="text/css" href="../static/siteinfo.css" />
+    <link rel="stylesheet" type="text/css" href="{{ url_for('static', path='/siteinfo.css')}}" />
 </head>
 
 <body>
